+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 Format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2025-04-05
+
+### Added
+- Public `get_term_properties(term_id)` method on `GraphBackend` interface,
+  implemented in both CogDB and Neo4j backends. Consumers no longer need to
+  reach into the private `_get_props()` helper to access `dcat_theme` and
+  other term attributes.
+
 ## [0.1.1] - 2025-04-04
 
 ### Added